# apps/backend/design_docs/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: ACMG-PS3 Intelligent Parsing API
  description: |-
    Versioned REST contract for the FastAPI backend that powers the ACMG-PS3
    multilingual document evidence extraction platform. The specification mirrors
    the current `/api/v1` implementation (FastAPI routers in
    `src/presentation/controllers`) so that the generated Swagger UI stays in sync
    with the running service.

    All endpoints are asynchronous-friendly. Uploads immediately return a task
    identifier, while task state can be polled via REST or streamed through the
    `/ws/task/{task_id}/progress` WebSocket.
  version: 2.1.0
  contact:
    name: ACMG-PS3 Engineering
    email: support@example.com
servers:
  - url: http://localhost:8000
    description: FastAPI root (dev)
  - url: http://localhost:8000/api/v1
    description: Versioned REST prefix
...
paths:
...
  /ws/task/{task_id}/progress:
    parameters:
      - $ref: '#/components/parameters/TaskIdPath'
    get:
      tags: [Realtime Progress]
      summary: Upgrade to WebSocket for live progress updates
      description: >
        Upgrades the HTTP connection to a WebSocket (managed by
        `src/presentation/websocket/progress_handler.ProgressHandler`) and
        streams JSON progress events published by Redis.
      operationId: progressWebSocket
      responses:
        '101':
          description: WebSocket handshake complete.
        '404':
          description: No handler available (route missing in deployment).
components:
  parameters:
    TaskIdPath:
      name: task_id
      in: path
      required: true
      schema:
        type: string
        description: UUID of the parsing task.
...